Big structure refactor, glfw deleted

+++ tools/windows/buildEditorWindows.bat
@@ -14,7 +14,9 @@ set CommonLinkerFlags=-incremental:no -opt:ref user32.lib gdi32.lib kernel32.lib
 
 pushd Envari
 pushd source
+pushd Scripting
 for /f "delims=" %%i in ('"forfiles /m LUAScriptingBindings.cpp /c "cmd /c echo @ftime" "') do set ScriptingDate=%%i
+popd
 popd
 popd
 
@@ -40,17 +42,35 @@ REM %random% to get random number
 @echo Start time %time%
 if not exist LUAScriptingBindings.%ScriptingDate::=.%.tmp (
     @echo Rebuilding LUA Bindings
-    cl -c ..\..\Envari\source\LUAScriptingBindings.cpp -FmLUAScriptingBindings.map %CommonCompilerFlags% -Bt -I ..\..\Envari\SDL2\include -I ..\..\Envari\Optick\include -I ..\..\Envari\LUA\include
+    cl -c ..\..\Envari\source\Scripting\LUAScriptingBindings.cpp ^
+    -FmLUAScriptingBindings.map ^
+    %CommonCompilerFlags% ^ 
+    -Bt ^
+    -I ..\..\Envari\LUA\include ^
+    -I ..\..\Envari\source\Defines ^
+    -I ..\..\Envari\source\Engine ^
+    -I ..\..\Envari\STB ^
+    -I ..\..\Envari\Miniaudio ^
+    -I ..\..\Envari\Engine ^
+    -LIBPATH:"..\..\Envari\LUA\windows\lib\x64" ^
     del /F *.tmp >NUL 2>NUL
     echo timestamp > LUAScriptingBindings.%ScriptingDate::=.%.tmp
 )
 
-cl ..\..\Envari\source\EditorWindows.cpp LUAScriptingBindings.obj ^
-    -FmEditorWindows.map -PDB:EditorWindows.pdb ^
-    -Bt %CommonCompilerFlags% ^
+cl ..\..\Envari\source\Runtimes\EditorWindows.cpp LUAScriptingBindings.obj ^
+    -FmEditorWindows.map -PDB:EditorWindows.pdb -Bt ^
+    %CommonCompilerFlags% ^
     -I ..\..\Envari\SDL2\include ^
-    -I ..\..\Envari\Optick\include ^
     -I ..\..\Envari\LUA\include ^
+    -I ..\..\Envari\source\Engine ^
+    -I ..\..\Envari\source\Defines ^
+    -I ..\..\Envari\source\Tools ^
+    -I ..\..\Envari\source\Default ^
+    -I ..\..\Envari\STB ^
+    -I ..\..\Envari\GL3W ^
+    -I ..\..\Envari\IMGUI ^
+    -I ..\..\Envari\Miniaudio ^
+    -I ..\..\Envari\ZSTD ^
     -link %CommonLinkerFlags% ^
     -LIBPATH:"..\..\Envari\SDL2\windows\x64" ^
     -LIBPATH:"..\..\Envari\Optick\lib\x64\release" ^
