+++ tools/windows/buildWindows64.bat
@@ -12,7 +12,9 @@ set CommonLinkerFlags=-incremental:no -opt:ref user32.lib gdi32.lib kernel32.lib
 
 pushd Envari
 pushd source
+pushd Scripting
 for /f "delims=" %%i in ('"forfiles /m LUAScriptingBindings.cpp /c "cmd /c echo @ftime" "') do set ScriptingDate=%%i
+popd
 popd
 popd
 
@@ -31,17 +33,36 @@ del /F *.pdb >NUL 2>NUL
 @echo Start time %time%
 if not exist LUAScriptingBindings.%ScriptingDate::=.%.tmp (
     @echo Rebuilding LUA Bindings
-    cl -c ..\..\Envari\source\LUAScriptingBindings.cpp -FmLUAScriptingBindings.map %CommonCompilerFlags% -Bt -I ..\..\Envari\SDL2\include -I ..\..\Envari\LUA\include
+    cl -c ..\..\Envari\source\Scripting\LUAScriptingBindings.cpp ^
+    -FmLUAScriptingBindings.map ^
+    %CommonCompilerFlags% ^ 
+    -Bt ^
+    -I ..\..\Envari\LUA\include ^
+    -I ..\..\Envari\source\Defines ^
+    -I ..\..\Envari\source\Engine ^
+    -I ..\..\Envari\STB ^
+    -I ..\..\Envari\Miniaudio ^
+    -I ..\..\Envari\Engine ^
+    -LIBPATH:"..\..\Envari\LUA\windows\lib\x64" ^
     del /F *.tmp >NUL 2>NUL
     echo timestamp > LUAScriptingBindings.%ScriptingDate::=.%.tmp
 )
 
-cl ..\..\Envari\source\RuntimeWindows.cpp LUAScriptingBindings.obj ^
+cl ..\..\Envari\source\Runtimes\RuntimeWindows.cpp LUAScriptingBindings.obj ^
     -FmRuntimeWindows.map %CommonCompilerFlags% -Bt ^
-    -I ..\..\Envari\SDL2\include -I ..\..\Envari\LUA\include ^
+    -I ..\..\Envari\LUA\include ^
+    -I ..\..\Envari\source\Engine ^
+    -I ..\..\Envari\source\Defines ^
+    -I ..\..\Envari\source\Tools ^
+    -I ..\..\Envari\source\Default ^
+    -I ..\..\Envari\STB ^
+    -I ..\..\Envari\GL3W ^
+    -I ..\..\Envari\IMGUI ^
+    -I ..\..\Envari\Miniaudio ^
+    -I ..\..\Envari\ZSTD ^
     -link %CommonLinkerFlags% ^
-    -LIBPATH:"..\..\Envari\SDL2\lib\x64" -LIBPATH:"..\..\Envari\LUA\windows\lib\x64" ^
-    -PDB:RuntimeWindows.pdb
+    -LIBPATH:"..\..\Envari\SDL2\lib\x64" ^
+    -LIBPATH:"..\..\Envari\LUA\windows\lib\x64" ^
 @echo End time %time%
 
 popd
